chore: Update fastMetaToEvent map in Index struct
This commit updates the `Index` struct in the `index.go` file by adding a new `fastMetaToEvent` map and its corresponding lock. This map is used to store the mapping between a string representation of `FastMetaParameter` and a slice of `types.Hash`. The `fastMetaToEvent` map is populated in the `RebuildFastMeta` method and is used in the `GetEventHashesByFastMetaParameter` and `GetEventHashesByFastMeta` methods to retrieve event hashes based on fast meta parameters. The changes also include the addition of the `GetEventsByFastMeta` method, which retrieves events based on a given `FastMeta` parameter.

```