chore: Update GarbageCollectionInterval check in NewOuroborosDB
The code change updates the check for the GarbageCollectionInterval in the NewOuroborosDB function. Instead of checking if the interval is zero or negative, it now checks if the interval is smaller than 2 minutes. If it is, a warning message is logged and the interval is set to the default value of 5 minutes.